[MCH] code cleanup
- Removed unnecessary headers
- Moved some header inclusions into source files
- Moved some template definitions into source files
- Removed obsolete class members
