+++ Modules/MUON/MCH/include/MCH/DecodingPostProcessing.h
@@ -19,8 +19,6 @@
 #ifndef QC_MODULE_MCH_PP_DIAGNOSTICS_H
 #define QC_MODULE_MCH_PP_DIAGNOSTICS_H
 
-#include "QualityControl/PostProcessingInterface.h"
-
 #include "MCH/PostProcessingConfigMCH.h"
 #include "MCH/Helpers.h"
 #include "Common/TH2Ratio.h"
@@ -28,6 +26,7 @@
 #include "MCH/DecodingErrorsPlotter.h"
 #include "MCH/HeartBeatPacketsPlotter.h"
 #include "MCH/FECSyncStatusPlotter.h"
+#include "QualityControl/PostProcessingInterface.h"
 
 #include <memory>
 
@@ -74,7 +73,6 @@ class DecodingPostProcessing : public PostProcessingInterface
   static std::string hbPacketsSourceName() { return "hbpackets"; }
   static std::string syncStatusSourceName() { return "syncstatus"; }
 
-  int64_t mRefTimeStamp;
   bool mFullHistos{ false };
 
   PostProcessingConfigMCH mConfig;
