+++ Modules/MUON/MCH/include/MCH/PedestalsCheck.h
@@ -19,10 +19,7 @@
 
 #include "MCH/Helpers.h"
 #include "QualityControl/CheckInterface.h"
-#include "QualityControl/MonitorObject.h"
 #include "QualityControl/Quality.h"
-#include "MCHRawCommon/DataFormats.h"
-#include "MCHRawElecMap/Mapper.h"
 
 namespace o2::quality_control_modules::muonchambers
 {
